Crud de Activos

+++ app/Http/Controllers/RecursoController.php
@@ -8,9 +8,8 @@ use Illuminate\Support\Facades\DB;
 use Illuminate\Support\Facades\Auth;  
 use App\audit;
 use App\Recurso;
-use App\SolicitudServicio;
 use App\cliente;
-use App\ResiduosGener;
+use App\SolicitudResiduo;
 
 class RecursoController extends Controller
 {
@@ -19,19 +18,14 @@ class RecursoController extends Controller
      *
      * @return \Illuminate\Http\Response
      */
-    public function index(){
-
-        $Recursos2 = DB::table('recursos')
-            ->join('residuos_geners', 'residuos_geners.ID_SGenerRes', '=', 'recursos.FK_ResGer')
-            ->select('recursos.*')
+    public function index()
+    {
+        $Recursos = DB::table('solicitud_residuos')
+            ->join('respels', 'respels.ID_Respel', '=', 'solicitud_residuos.FK_SolResRespel')
+            ->select('respels.RespelName', 'solicitud_residuos.FK_SolResSolSer', 'solicitud_residuos.SolResSlug')
             ->get();
 
-        $Recursos = DB::table('residuos_geners')
-            ->join('respels', 'respels.ID_Respel', '=', 'residuos_geners.FK_Respel')
-            ->select('respels.RespelName', 'residuos_geners.*')
-            ->get();
-
-        return view('recursos.index', compact('Recursos', 'Recursos2'));
+        return view('recursos.index', compact('Recursos'));
     }
 
     /**
@@ -43,18 +37,12 @@ class RecursoController extends Controller
     {
         $Clientes = cliente::all();
 
-        $ResGeners = DB::table('residuos_geners')
-            ->join('respels', 'respels.ID_Respel', '=', 'residuos_geners.FK_Respel')
-            ->select('respels.RespelName', 'residuos_geners.FK_SolSer', 'residuos_geners.ID_SGenerRes')
-            ->get();
+        $SolRes = DB::table('solicitud_residuos')
+        ->join('respels', 'respels.ID_Respel', '=', 'solicitud_residuos.FK_SolResRespel')
+        ->select('respels.RespelName', 'solicitud_residuos.FK_SolResSolSer', 'solicitud_residuos.ID_SolRes')
+        ->get();
 
-        $Recursos = DB::table('recursos')
-            ->join('residuos_geners', 'residuos_geners.ID_SGenerRes', '=', 'recursos.FK_ResGer')
-            ->join('solicitud_servicios', 'solicitud_servicios.ID_SolSer', '=', 'residuos_geners.FK_SolSer')
-            ->select('recursos.*', 'residuos_geners.FK_SolSer')
-            ->get();
-
-        return view('recursos.create', compact('ResGeners', 'Clientes', 'Recursos'));
+        return view('recursos.create', compact('SolRes', 'Clientes', 'Recursos'));
     }
 
     /**
@@ -65,7 +53,7 @@ class RecursoController extends Controller
      */
     public function store(Request $request)
     {
-        if ($request->hasfile('RecSrc')) {
+        if ($request->hasfile('RecSrc')){
             foreach($request->RecSrc as $file){ 
             
             $Recurso = new Recurso();
@@ -82,7 +70,8 @@ class RecursoController extends Controller
             $Recurso->SlugRec = 'Slug'.$name;
             $Recurso->RecSrc = $Src;
             $Recurso->RecFormat = '.'.$Extension;
-            $Recurso->FK_ResGer = $request->input("FK_ResGer");
+            $Recurso->RecDelete = 0;
+            $Recurso->FK_RecSolRes = $request->input("FK_RecSolRes");
             $Recurso->save();
             }
         }
@@ -97,15 +86,15 @@ class RecursoController extends Controller
      */
     public function show(Request $request, $id)
     {
-        $ResGeners = ResiduosGener::where('ID_SGenerRes', $id)->first();
-        
+        $SolRes = SolicitudResiduo::where('SolResSlug', $id)->first();
+
         $Recursos = DB::table('recursos')
-            ->join('residuos_geners', 'residuos_geners.ID_SGenerRes', '=', 'recursos.FK_ResGer')
-            ->select('recursos.*', 'residuos_geners.ID_SGenerRes')
-            ->where('FK_ResGer', $id)
+            ->join('solicitud_residuos', 'solicitud_residuos.ID_SolRes', '=', 'recursos.FK_RecSolRes')
+            ->select('recursos.*', 'solicitud_residuos.SolResSlug')
+            ->where('FK_RecSolRes', $SolRes->ID_SolRes)
             ->get();
 
-        return view('recursos.show', compact('Recursos', 'ResGeners'));
+        return view('recursos.show', compact('Recursos', 'SolRes'));
     }
     
     /**
@@ -116,20 +105,23 @@ class RecursoController extends Controller
      */
     public function edit($id)
     {
-        $ResGeners = ResiduosGener::where('ID_SGenerRes', $id)->first();
+        $SolResiduos = DB::table('solicitud_residuos')
+        ->join('respels', 'respels.ID_Respel', '=', 'solicitud_residuos.FK_SolResRespel')
+        ->select('respels.RespelName', 'solicitud_residuos.ID_SolRes', 'solicitud_residuos.FK_SolResSolSer')
+        ->get();
+
+        $SolRes = SolicitudResiduo::where('SolResSlug', $id)->first();
+        $Recs = Recurso::where('FK_RecSolRes', $SolRes->ID_SolRes)->first();
 
         $Clientes = cliente::all();
 
-        $SolServs = SolicitudServicio::all();
-
-        // $SolServs = DB::table('solicitud_servicios')
-        // ->join('residuos_geners', 'solicitud_servicios.ID_SolSer', '=', 'residuos_geners.FK_Respel')
-        // ->join('respels', 'respels.ID_Respel', '=', 'residuos_geners.FK_Respel')
-        // ->select('respels.RespelName', 'solicitud_servicios.*')
-        // // ->where('residuos_geners.ID_SGenerRes', $id)
-        // ->get();
-
-        return view('recursos.edit', compact('ResGeners', 'Clientes', 'SolServs'));
+        $Recursos = DB::table('recursos')
+            ->join('solicitud_residuos', 'solicitud_residuos.ID_SolRes', '=', 'recursos.FK_RecSolRes')
+            ->select('recursos.*', 'solicitud_residuos.SolResSlug', 'solicitud_residuos.FK_SolResSolSer', 'solicitud_residuos.ID_SolRes')
+            ->where('FK_RecSolRes',  $SolRes->ID_SolRes)
+            ->get();
+        
+        return view('recursos.edit', compact('SolResiduos', 'Clientes', 'SolRes', 'Recursos', 'Recs'));
     }
 
     /**
@@ -141,27 +133,35 @@ class RecursoController extends Controller
      */
     public function update(Request $request, $id)
     {
+        // edit
         if($request->input("number") == 0){
-            $Recursos = Recurso::where('FK_ResGer', $id)->first();
+            $SolRes = SolicitudResiduo::where('SolResSlug', $id)->first();
+            $Recursos = Recurso::where('FK_RecSolRes', $SolRes->ID_SolRes)->first();
+            
+            // return $SolRes;
+            // modificar el nombre de la carpeta
             rename(public_path($Recursos->RecSrc), 'Recursos/'.$request->input("RecName").time());
 
-            $Recurso = Recurso::where('FK_ResGer', $id)->update(['RecName' => $request->input("RecName") ,'RecSrc' => 'Recursos/'.$request->input("RecName").time()]);
+            $Recurso = Recurso::where('FK_RecSolRes',$SolRes->ID_SolRes)->update([
+                    'RecName' => $request->input("RecName"),
+                    'RecSrc' => 'Recursos/'.$request->input("RecName").time(), 
+                    'FK_RecSolRes' => $request->input("FK_RecSolRes")
+                ]);
 
-            $ResGeners = ResiduosGener::where('ID_SGenerRes', $id)->update(['FK_SolSer' => $request->input("FK_SolSer")]);
-            
             $log = new audit();
-            $log->AuditTabla="residuos_geners y recursos";
+            $log->AuditTabla="recursos";
             $log->AuditType="Modificado";
-            $log->AuditRegistro = $ResGeners->ID_SGenerRes;
+            $log->AuditRegistro = $Recursos->FK_RecSolRes;
             $log->AuditUser=Auth::user()->email;
             $log->Auditlog=$request->all();
             $log->save();
 
         // return redirect()->route('recurso.index');
-
         }
+        // modal en show
         if($request->input("number") == 1){
-            $Recursos = Recurso::where('FK_ResGer', $id)->first();
+            $SolRes = SolicitudResiduo::where('SolResSlug', $id)->first();
+            $Recursos = Recurso::where('FK_RecSolRes', $SolRes->ID_SolRes)->first();
 
             if ($request->hasfile('RecSrc')) {
                 foreach($request->RecSrc as $file){ 
@@ -180,7 +180,8 @@ class RecursoController extends Controller
                 $Recurso->SlugRec = 'Slug'.$name;
                 $Recurso->RecSrc = $Recursos->RecSrc;
                 $Recurso->RecFormat = '.'.$Extension;
-                $Recurso->FK_ResGer = $Recursos->FK_ResGer;
+                $Recurso->FK_RecSolRes = $Recursos->FK_RecSolRes;
+                $Recurso->RecDelete = $Recursos->RecDelete;
                 $Recurso->save();
                 }
             }
@@ -195,23 +196,40 @@ class RecursoController extends Controller
      * @param  int  $id
      * @return \Illuminate\Http\Response
      */
-    public function destroy($id)
-    {
-        $Recursos = Recurso::where('ID_Rec', $id)->first();
+    public function destroy(Request $request, $id)
+    {   
+        if($request->input("number") == 0){
+            $SolRes = SolicitudResiduo::where('SolResSlug', $id)->first();
+            $Recs = Recurso::where('FK_RecSolRes', $SolRes->ID_SolRes)->first();
+    
+            $Recursos = DB::table('recursos')
+                ->join('solicitud_residuos', 'solicitud_residuos.ID_SolRes', '=', 'recursos.FK_RecSolRes')
+                ->select('recursos.*', 'solicitud_residuos.SolResSlug', 'solicitud_residuos.FK_SolResSolSer', 'solicitud_residuos.ID_SolRes')
+                ->where('FK_RecSolRes',  $SolRes->ID_SolRes)
+                ->get();
 
-        unlink(public_path($Recursos->RecSrc)."/".$Recursos->RecRmSrc);
+            foreach($Recursos as $Recurso){
+                if ($Recurso->RecDelete == 0){
+                    $Rec = Recurso::where('FK_RecSolRes',$SolRes->ID_SolRes)->update(['RecDelete' => 1]);
+                }
+                else{
+                    $Rec = Recurso::where('FK_RecSolRes',$SolRes->ID_SolRes)->update(['RecDelete' => 0]);
+                }
+            }        
+            $log = new audit();
+            $log->AuditTabla = "recursos";
+            $log->AuditType = "Eliminado";
+            $log->AuditRegistro = $Recs->FK_RecSolRes;
+            $log->AuditUser = Auth::user()->email;
+            $log->Auditlog = $Rec;
+            $log->save();
+        }
 
-        Recurso::destroy($id);
-
-
-        // $log = new audit();
-        // $log->AuditTabla = "recursos";
-        // $log->AuditType = "Eliminado";
-        // $log->AuditRegistro = $Recursos->ID_Rec;
-        // $log->AuditUser = Auth::user()->email;
-        // $log->Auditlog = $Persona->PersDelete;
-        // $log->save();
-
+        if($request->input("number") == 1){
+            $Recursos = Recurso::where('ID_Rec', $id)->first();
+            unlink(public_path($Recursos->RecSrc)."/".$Recursos->RecRmSrc);
+            Recurso::destroy($id);
+        }
         return redirect()->route('recurso.index');
     }
 }
